Update generate
After a second thought it's better to have generate and similar commands
run against each project.

The reasoning is that having the command only here would be confusing,
as the playground command would have a 'state' hard to see and to
remember.
With every command having their own generate will make it easier to
maintain and to remove if necessary